New LazyWebElement that takes a WebElement supplier
